Navigating flow from page to page works, but there are two warning messages for missing dependencies on the useEffect

+++ client/src/Pages/CreateRide/CreateRide.js
@@ -1,5 +1,6 @@
 import React from 'react';
-import Create from './Create.js'; 
+import Create from './Create.js';
+import { useEffect } from 'react'; 
 import { useHistory } from "react-router";
 import { gql, useMutation } from "@apollo/client";
 import { useToasts } from "react-toast-notifications";
@@ -8,8 +9,8 @@ const CreateRide = () => {
 
     const history = useHistory();
     const { addToast } = useToasts();
-    // Set the last page visited
-    localStorage.setItem('lastPage', 'create-ride');
+    const previous = localStorage.getItem('from');
+
     const CREATE_RIDE = gql`
         mutation CreateRide (
             $startLoc: MongoID!, $endLoc: MongoID!, $date: Date, $passengers: Float, $users: [MongoID!], $owner: MongoID!) 
@@ -48,19 +49,32 @@ const CreateRide = () => {
         });
     }
 
-    // Back Button Logic (in case of OnBoarding Prompt)
-    // TODO: This is supposed to display a confirmation message if the user attempts to "go back", 
-    // but for some reason it doesn't quite work. 
-    function beforeUnloadListener(event) {
-        event.preventDefault();
-        console.log('back button detected');
-        localStorage.setItem('from', 'createRide'); 
-            history.goBack();
-    };
+    // In case of returning to onboarding prompt, 
+    // the user needs to be warned that navigating to search via navbar is the expected action. 
+    const onBackButtonEvent = (e) => {
+        e.preventDefault();
+        if (previous === 'onboardingCompletion'){
+            if (window.confirm("Do you want to navigate back to onboarding form? It is advised to navigate back to the search page with the sidebar.")) {
+                // Clear the localStorage to ensure that the user starts on a fresh profile
+                localStorage.clear();
+                localStorage.setItem('from', 'createRide');
+                history.goBack();
+            } else {
+                window.history.pushState(null, null, window.location.pathname);
+            }
+        }
+        
+      }
 
-    // Include safety net for when user accidentally tries to go back
-    window.addEventListener("beforeunload", beforeUnloadListener);
-
+    // Track back button behavior
+    useEffect(() => {
+        window.history.pushState(null, null, window.location.pathname);
+        window.addEventListener('popstate', onBackButtonEvent);
+        return () => {
+            window.removeEventListener('popstate', onBackButtonEvent);  
+        };
+    }, []);
+    
 
     return (
